+++ CHANGELOG.md
@@ -6,3 +6,5 @@ All notable changes to the "solargraph" extension will be documented in this fil
 
 ## 0.0.2
 - Updated README and package metadata (author, keywords, etc.)
+- Added icon
+- Reduced console output